Make card page component
* card page is editing items to preview card.
* card_gen is generate header at fetched from card validator.
